Dependency updates and related changes.

+++ app/src/app/consts.rs
@@ -7,4 +7,4 @@ use treeview::SidebarPosition;
 pub use treeview::TXT_SIZE;
 
 pub(crate) const SIDEBAR_POSITION: SidebarPosition = SidebarPosition::Right;
-pub(crate) const APP_SCALE_FACTOR: f64 = 1e0 / SF as f64;
+pub(crate) const APP_SCALE_FACTOR: f32 = 1e0 / SF;
